fix: Restore the offset 1 of BOP . Remove SetAssociativeRegs style: and rename Queue_Regs to OverwriteQueue
